+++ src/plugins/relay/relay-buffer.c
@@ -79,15 +79,15 @@ relay_buffer_refresh (const char *hotlist)
                       /* disconnect */
                       (client_selected
                        && !RELAY_STATUS_HAS_ENDED(client_selected->status)) ?
-                      _("  [D] Disconnect") : "",
+                      _("  [d] Disconnect") : "",
                       /* remove */
                       (client_selected
                        && RELAY_STATUS_HAS_ENDED(client_selected->status)) ?
-                      _("  [R] Remove") : "",
+                      _("  [r] Remove") : "",
                       /* purge old */
-                      _("  [P] Purge finished"),
+                      _("  [p] Purge finished"),
                       /* quit */
-                      _("  [Q] Close this buffer"));
+                      _("  [q] Close this buffer"));
     for (ptr_client = relay_clients; ptr_client;
          ptr_client = ptr_client->next_client)
     {
@@ -160,7 +160,7 @@ relay_buffer_refresh (const char *hotlist)
         line++;
     }
     if (hotlist)
-        weechat_buffer_set (relay_buffer, "hotlist", hotlist);
+        weechat_buffer_set (relay_buffer, "hotlist_conditions", hotlist);
 }
 
 /*
@@ -174,6 +174,7 @@ relay_buffer_input_cb (const void *pointer, void *data,
 {
     struct t_relay_client *client, *ptr_client, *next_client;
     const char *ptr_remote_name, *ptr_remote_id;
+    int refresh;
 
     /* make C compiler happy */
     (void) pointer;
@@ -200,15 +201,20 @@ relay_buffer_input_cb (const void *pointer, void *data,
         /* purge old clients */
         else if (weechat_strcmp (input_data, "p") == 0)
         {
+            refresh = 0;
             ptr_client = relay_clients;
             while (ptr_client)
             {
                 next_client = ptr_client->next_client;
                 if (RELAY_STATUS_HAS_ENDED(ptr_client->status))
+                {
                     relay_client_free (ptr_client);
+                    refresh = 1;
+                }
                 ptr_client = next_client;
             }
-            relay_buffer_refresh (WEECHAT_HOTLIST_MESSAGE);
+            if (refresh)
+                relay_buffer_refresh (WEECHAT_HOTLIST_MESSAGE);
         }
         /* quit relay buffer (close it) */
         else if (weechat_strcmp (input_data, "q") == 0)