+++ tests/unit/plugins/relay/test-relay-http.cpp
@@ -541,6 +541,24 @@ TEST(RelayHttp, ParseHeader)
     request = relay_http_request_alloc ();
     CHECK(request);
     relay_http_parse_method_path (request, "GET /api/version");
+    relay_http_parse_header (request, "Accept-Encoding: gzip", 1);
+    relay_http_parse_header (request, "Accept-Encoding: zstd", 1);
+    LONGS_EQUAL(RELAY_HTTP_HEADERS, request->status);
+    STRCMP_EQUAL("GET /api/version\n"
+                 "Accept-Encoding: gzip\n"
+                 "Accept-Encoding: zstd\n",
+                 *(request->raw));
+    LONGS_EQUAL(1, request->headers->items_count);
+    STRCMP_EQUAL("gzip, zstd",
+                 (const char *)hashtable_get (request->headers, "accept-encoding"));
+    LONGS_EQUAL(2, request->accept_encoding->items_count);
+    CHECK(hashtable_has_key (request->accept_encoding, "gzip"));
+    CHECK(hashtable_has_key (request->accept_encoding, "zstd"));
+    relay_http_request_free (request);
+
+    request = relay_http_request_alloc ();
+    CHECK(request);
+    relay_http_parse_method_path (request, "GET /api/version");
     relay_http_parse_header (request, "Content-Length: 123", 1);
     LONGS_EQUAL(RELAY_HTTP_HEADERS, request->status);
     STRCMP_EQUAL("GET /api/version\n"
@@ -866,6 +884,27 @@ TEST(RelayHttp, GetAuthStatus)
     free (totp2);
     config_file_option_reset (relay_config_network_totp_secret, 1);
     config_file_option_reset (relay_config_network_totp_window, 1);
+
+    /* test invalid plain-text password ("test") via Sec-WebSocket-Protocol */
+    hashtable_remove (client->http_req->headers, "authorization");
+    hashtable_set (client->http_req->headers, "sec-websocket-protocol",
+                   WEBSOCKET_SUB_PROTOCOL_API_WEECHAT
+                   ", base64url.bearer.authorization.weechat.cGxhaW46dGVzdA");
+    LONGS_EQUAL(-2, relay_http_get_auth_status (client));
+
+    /* test valid plain-text password ("secret_password") via Sec-WebSocket-Protocol */
+    hashtable_set (client->http_req->headers, "sec-websocket-protocol",
+                   WEBSOCKET_SUB_PROTOCOL_API_WEECHAT
+                   ", base64url.bearer.authorization.weechat.cGxhaW46c2VjcmV0X3Bhc3N3b3Jk");
+    LONGS_EQUAL(0, relay_http_get_auth_status (client));
+
+    /* test auth via Sec-WebSocket-Protocol with base64url specific characters */
+    config_file_option_set (relay_config_network_password, "..>..?.", 1);
+    hashtable_set (client->http_req->headers, "sec-websocket-protocol",
+                   WEBSOCKET_SUB_PROTOCOL_API_WEECHAT
+                   ", base64url.bearer.authorization.weechat.cGxhaW46Li4-Li4_Lg");
+    LONGS_EQUAL(0, relay_http_get_auth_status (client));
+    config_file_option_set (relay_config_network_password, good_pwd, 1);
 
     config_file_option_reset (relay_config_network_password, 1);